refactor: move snapshot and snapshot-check logic to gas-snapshots module
